requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy<2.0.0
torch==2.1.2
torchvision==0.16.2
easyocr==1.7.1
opencv-python==4.9.0.80
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-decouple==3.8
